[FPGA] Fix global size of work-group autorun kernels
Work-group autorun kernels should be launched with global_size = (2^32,
2^32, 2^32). Fixed set of issues related to unsigned integer
overflow while using 2^32 as value for global_size.

Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/129751
Tested-by: Sachkov, Alexey <alexey.sachkov@intel.com>

+++ opencl-intel/cl_api/task_executor.h
@@ -338,7 +338,7 @@ public:
 
     // Optimize By
     virtual TASK_SET_OPTIMIZATION OptimizeBy()                        const = 0;
-    virtual unsigned int          PreferredSequentialItemsPerThread() const = 0;
+    virtual size_t PreferredSequentialItemsPerThread() const = 0;
 
 };
